refactor: creat multisig util methods for timelocks
